test: AnswerEntrySet のテスト配置を実装に合わせる
AnswerEntrySet へ移した認可ロジックのテストが空の answer::service モジュールに残らないよう、実装側へ移動する。
