+++ server/domain/src/form/answer.rs
@@ -1,3 +1,2 @@
 pub mod models;
-pub mod service;
 pub mod settings;
